backend/pyproject.toml
[project]
name = "user-twin-studio-backend"
version = "0.1.0"
description = "Local-first backend for distilling a user operating system skill."
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
  "pypdf",
  "olefile",
  "certifi",
  "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
